VoteOnSuspectNodeHelpe -> ContractFacade unit test
